requirements/base.txt
Django
psycopg2-binary
python-dateutil
Faker